Hotfix: Aplicacion de permisos Socios

+++ app/DataTables/SociosDataTable.php
@@ -4,6 +4,7 @@ namespace App\DataTables;
 
 use App\Helpers\Helpers;
 use App\Models\Socios;
+use App\Providers\PermissionKey;
 use Illuminate\Database\Eloquent\Builder as QueryBuilder;
 use Yajra\DataTables\EloquentDataTable;
 use Yajra\DataTables\Html\Builder as HtmlBuilder;
@@ -25,32 +26,55 @@ class SociosDataTable extends DataTable
 		return (new EloquentDataTable($query))
 			->setRowId('id')
 			->addColumn('acciones', function (Socios $socio) {
-				return
-					'
-				<a title="Editar socio" href="' . route("panel.socios.edit", ["id" => $socio->id]) . '" class="text-white bg-yellow-700 hover:bg-yellow-800 font-medium rounded-lg text-xs px-2 py-1 mr-2 inline-flex items-center justify-center" style="width: 36px;">
-					<svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-edit w-[20px] inline" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
-						<path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-						<path d="M7 7h-1a2 2 0 0 0 -2 2v9a2 2 0 0 0 2 2h9a2 2 0 0 0 2 -2v-1"></path>
-						<path d="M20.385 6.585a2.1 2.1 0 0 0 -2.97 -2.97l-8.415 8.385v3h3l8.385 -8.415z"></path>
-						<path d="M16 5l3 3"></path>
-					</svg>
-				</a>
-				<form action="' . route("panel.socios.destroy", ["id" => $socio->id]) . '" method="post" class="inline delete-form-' . $socio->id . '">
-					<input type="hidden" name="_token" value="' . csrf_token() . '" />
-					<input type="hidden" name="_method" value="DELETE">
-					<button title="Eliminar" type="button" onclick="deleteSubmitForm(' . $socio->id . ')" class="text-white bg-red-700 hover:bg-red-800 font-medium rounded-lg text-xs px-2 py-1 mr-2 inline-flex items-center justify-center" style="width: 36px">
-						<svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-file-info w-[20px] inline" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
-							<path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-							<path d="M4 7l16 0"></path>
-							<path d="M10 11l0 6"></path>
-							<path d="M14 11l0 6"></path>
-							<path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12"></path>
-							<path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3"></path>
-						</svg>
-					</button>
-				</form>
-				
-				';
+
+				$acciones = '';
+
+				if (auth()->user()->hasAllDirectPermissions([PermissionKey::Socios['permissions']['edit']['name'], PermissionKey::Socios['permissions']['update']['name']])) {
+					$acciones .= '
+						<a  href="' . route("panel.socios.edit", ["id" => $socio->id]) . '" class="font-medium text-orange-600 dark:text-orange-500 " title="Editar">
+							<svg class="w-5 inline" aria-hidden="true"
+								fill="none" stroke="currentColor" stroke-width="1.5" viewBox="0 0 24 24"
+								xmlns="http://www.w3.org/2000/svg">
+								<path
+									d="M16.862 4.487l1.687-1.688a1.875 1.875 0 112.652 2.652L10.582 16.07a4.5 4.5 0 01-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 011.13-1.897l8.932-8.931zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0115.75 21H5.25A2.25 2.25 0 013 18.75V8.25A2.25 2.25 0 015.25 6H10"
+									stroke-linecap="round" stroke-linejoin="round"></path>
+							</svg>
+						</a>';
+				} else {
+					if (auth()->user()->hasDirectPermission(PermissionKey::Socios['permissions']['edit']['name'])) {
+						$acciones .= '
+							<a href="' . route("panel.socios.edit", ["id" => $socio->id]) . '" class="font-medium text-blue-600 dark:text-blue-500 " title="Ver detalle">
+								<svg width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
+									<path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+									<path d="M3 19a9 9 0 0 1 9 0a9 9 0 0 1 9 0"></path>
+									<path d="M3 6a9 9 0 0 1 9 0a9 9 0 0 1 9 0"></path>
+									<path d="M3 6l0 13"></path>
+									<path d="M12 6l0 13"></path>
+									<path d="M21 6l0 13"></path>
+								</svg>
+							</a>';
+					}
+				}
+
+				if (auth()->user()->hasDirectPermission(PermissionKey::Socios['permissions']['destroy']['name'])) {
+					$acciones .= '
+						<form action="' . route("panel.socios.destroy", ["id" => $socio->id]) . '" method="post" class="inline delete-form-' . $socio->id . '">
+							<input type="hidden" name="_token" value="' . csrf_token() . '" />
+							<input type="hidden" name="_method" value="DELETE">
+							<button title="Eliminar" type="button" onclick="deleteSubmitForm(' . $socio->id . ')" class="text-white bg-red-700 hover:bg-red-800 font-medium rounded-lg text-xs px-2 py-1 mr-2 inline-flex items-center justify-center" style="width: 36px">
+								<svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-file-info w-[20px] inline" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
+									<path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+									<path d="M4 7l16 0"></path>
+									<path d="M10 11l0 6"></path>
+									<path d="M14 11l0 6"></path>
+									<path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12"></path>
+									<path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3"></path>
+								</svg>
+							</button>
+						</form>';
+				}
+
+				return $acciones;
 			})
 			->addColumn('fecha_inicio', function (Socios $socios) {
 				return Helpers::dateSpanishShort($socios->fecha_inicio);
